Created a styleguide document to maintain the styleguide of our projects
This doc should be used to document the coding style. Contributing styles should be maintained in the contributing doc. Id like to split the docs into the git stuff and code stuff so that we always can reuse the git stuff but keep the code stuff as detailed as possible